+++ src/main/java/ro/webdata/normalization/timespan/ro/TimeExpression.java
@@ -8,12 +8,12 @@ public class TimeExpression {
     private String separator = "\n";
     private String value = null;
     private String sanitizedValue = null;
-    private TreeSet<String> edgesValues = new TreeSet<>();
-    private TreeSet<String> normalizedValues = new TreeSet<>();
-    private ArrayList<String> types = new ArrayList<>();
+    private List<Map<String, String>> edgesValues = new ArrayList<>();
+    private Set<String> normalizedValues = new LinkedHashSet<>();
+    private List<String> types = new ArrayList<>();
 
     public static String getHeaders() {
-        ArrayList<String> headers = new ArrayList<>(){{
+        List<String> headers = new ArrayList<>(){{
             add("initial value");
             add("sanitized value");
             add("normalized values");
@@ -35,30 +35,11 @@ public class TimeExpression {
         this.sanitizedValue = TimeSanitizeUtils.sanitizeValue(value, null);
         TimespanModel timespanModel = TimespanUtils.prepareTimespanModel(this.sanitizedValue);
         this.normalizedValues = timespanModel.getDBpediaUris();
-        this.edgesValues = prepareEdgesValues(timespanModel);
+        this.edgesValues = timespanModel.getDBpediaEdgesUris();
         this.types = timespanModel.getTypes();
 
         if (separator != null)
             this.separator = separator;
-    }
-
-    private TreeSet<String> prepareEdgesValues(TimespanModel timespanModel) {
-        HashSet<String> edgesValues = new HashSet<>();
-        ArrayList<HashMap<String, String>> edgesEntries = timespanModel.getDBpediaEdgesUris();
-
-        for (HashMap<String, String> item : edgesEntries) {
-            for (Map.Entry<String, String> entry : item.entrySet()) {
-                String edge = entry.getValue();
-
-                if (edge != null)
-                    edgesValues.add(edge);
-            }
-        }
-
-        ArrayList<String> sorted = new ArrayList<>(edgesValues);
-        Collections.sort(sorted);
-
-        return new TreeSet<>(sorted);
     }
 
     @Override
@@ -78,7 +59,11 @@ public class TimeExpression {
         return sanitizedValue;
     }
 
-    public TreeSet<String> getNormalizedValues() {
+    public Set<String> getNormalizedValues() {
         return normalizedValues;
+    }
+
+    public List<Map<String, String>> getEdgesValues() {
+        return this.edgesValues;
     }
 }