+++ CHANGES.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 ---
 
+## 1.5
+- Bump commons-echo to 1.5
+- Bump LIDO-Parser to 1.2
+- Bump Dublin-Core-Parser to 1.2
+- Updated edgesValues field of TimeExpression to directly retrieve DBpedia URIs from timespanModel
+
 ## 1.4
 - Added gradle support